requirements.txt
beautifulsoup4==4.13.4
bleach==4.1.0
chromadb==0.6.3
cryptography==44.0.2
deep_translator==1.11.4
duckduckgo_search==8.0.0
fitz==0.0.1.dev2
Flask==3.1.0
graphviz==0.20.3
groq==0.23.0
langchain==0.3.24
langchain_community==0.3.22
langchain_core==0.3.55
langchain_openai==0.3.14
matplotlib==3.10.1
mistralai==1.7.0
numpy==2.2.5
ollama==0.4.8
openai==1.76.0
pandas==2.2.3
Pillow==11.2.1
plotly==5.24.1
pymupdf==1.25.5
PyPDF2==3.0.1
python-dotenv==1.1.0
requests==2.32.3
scikit_learn==1.6.1
seaborn==0.13.2
sentence_transformers==4.0.1
spacy==3.8.4
st_theme==1.2.3
streamlit==1.44.1
streamlit_extras==0.6.0
streamlit_float==0.3.5
streamlit_javascript==0.1.5
streamlit_option_menu==0.4.0
streamlit_theme==0.58.0
tika==2.6.0
tiktoken==0.8.0
torch==2.6.0
validators==0.34.0
wordcloud==1.9.4
